+++ tests/test_model_lock.py
@@ -53,7 +53,8 @@ class TestRfLock:
             parameter="STATE",
             value=True,
             wait_for_callback=WAIT_FOR_CALLBACK,
-            priority=CommandPriority.HIGH,
+            priority=CommandPriority.CRITICAL,
+            purge_addresses=frozenset({"VCU0000146:1"}),
         )
         assert lock.is_locked is False
         await lock.lock()
@@ -63,7 +64,8 @@ class TestRfLock:
             parameter="STATE",
             value=False,
             wait_for_callback=WAIT_FOR_CALLBACK,
-            priority=CommandPriority.HIGH,
+            priority=CommandPriority.CRITICAL,
+            purge_addresses=frozenset({"VCU0000146:1"}),
         )
         assert lock.is_locked is True
         await lock.open()
@@ -73,7 +75,8 @@ class TestRfLock:
             parameter="OPEN",
             value=True,
             wait_for_callback=WAIT_FOR_CALLBACK,
-            priority=CommandPriority.HIGH,
+            priority=CommandPriority.CRITICAL,
+            purge_addresses=frozenset({"VCU0000146:1"}),
         )
 
         assert lock.is_locking is None
@@ -150,7 +153,8 @@ class TestIpLock:
             parameter="LOCK_TARGET_LEVEL",
             value=_LockTargetLevel.LOCKED,
             wait_for_callback=WAIT_FOR_CALLBACK,
-            priority=CommandPriority.HIGH,
+            priority=CommandPriority.CRITICAL,
+            purge_addresses=frozenset({"VCU9724704:0", "VCU9724704:1"}),
         )
         await central.event_coordinator.data_point_event(
             interface_id=const.INTERFACE_ID, channel_address="VCU9724704:1", parameter="LOCK_STATE", value=1
@@ -163,7 +167,8 @@ class TestIpLock:
             parameter="LOCK_TARGET_LEVEL",
             value=_LockTargetLevel.UNLOCKED,
             wait_for_callback=WAIT_FOR_CALLBACK,
-            priority=CommandPriority.HIGH,
+            priority=CommandPriority.CRITICAL,
+            purge_addresses=frozenset({"VCU9724704:0", "VCU9724704:1"}),
         )
         await central.event_coordinator.data_point_event(
             interface_id=const.INTERFACE_ID, channel_address="VCU9724704:1", parameter="LOCK_STATE", value=2
@@ -176,7 +181,8 @@ class TestIpLock:
             parameter="LOCK_TARGET_LEVEL",
             value=_LockTargetLevel.OPEN,
             wait_for_callback=WAIT_FOR_CALLBACK,
-            priority=CommandPriority.HIGH,
+            priority=CommandPriority.CRITICAL,
+            purge_addresses=frozenset({"VCU9724704:0", "VCU9724704:1"}),
         )
 
         assert lock.is_locking is None